+++ src/main/java/gg/packetloss/grindstone/sacrifice/SacrificeResult.java
@@ -8,15 +8,16 @@ package gg.packetloss.grindstone.sacrifice;
 
 import org.bukkit.inventory.ItemStack;
 
+import java.math.BigDecimal;
 import java.util.Iterator;
 import java.util.List;
 
 public class SacrificeResult implements Iterable<ItemStack> {
     private List<ItemStack> itemStacks;
-    private double remainingValue;
+    private BigDecimal remainingValue;
     private int remainingItems;
 
-    public SacrificeResult(List<ItemStack> itemStacks, double remainingValue, int remainingItems) {
+    public SacrificeResult(List<ItemStack> itemStacks, BigDecimal remainingValue, int remainingItems) {
         this.itemStacks = itemStacks;
         this.remainingValue = remainingValue;
         this.remainingItems = remainingItems;
@@ -26,7 +27,7 @@ public class SacrificeResult implements Iterable<ItemStack> {
         return itemStacks;
     }
 
-    public double getRemainingValue() {
+    public BigDecimal getRemainingValue() {
         return remainingValue;
     }
